+++ src/interfaces/Field.ts
@@ -1,5 +1,5 @@
 export interface Field {
-  type: "text" | "email" | "file";
+  type: "text" | "email" | "file" | "date";
   value: string;
   onChange: (e: any) => void;
   label: string;