+++ src/pages/api/global/cad-info/features.ts
@@ -4,6 +4,8 @@ import { processQuery } from "@lib/database";
 import { logger } from "@lib/logger";
 import { IRequest } from "src/interfaces/IRequest";
 import { Cad } from "types/Cad";
+import { User } from "types/User";
+import useAuth from "@hooks/useAuth";
 
 export function parseFeatures(cad: Cad): string[] {
   let features: string[];
@@ -18,13 +20,28 @@ export function parseFeatures(cad: Cad): string[] {
 }
 
 export default async function (req: IRequest, res: NextApiResponse) {
+  try {
+    await useAuth(req);
+  } catch (e) {
+    null;
+  }
+
   switch (req.method) {
     case "GET": {
       try {
-        const [cad] = await processQuery<Cad>("SELECT `features` FROM `cad_info`");
+        const [user] = await processQuery<User>("SELECT `rank` FROM `users` WHERE `id` = ?", [
+          req.userId,
+        ]);
+        const [cad] = await processQuery<Cad>(
+          "SELECT `features`, `registration_code` FROM `cad_info`",
+        );
         const features = parseFeatures(cad!);
 
+        const code =
+          user?.rank === "owner" ? cad?.registration_code ?? "" : !!cad?.registration_code;
+
         return res.json({
+          registration_code: code,
           features,
           status: "success",
         });