+++ docs/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
+## 2.0.3
+
+- New: custom registration code (something like invite codes) ([#156](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/156))
+- New: admins can delete incidents
+- New: added new features to disable
+- New: added citizen images in /citizen page
+- New: added search input for /citizen page
+- Minor: fixed minor bugs
+- Minor: removed database timeouts
+- Minor: fixed small bug with status buttons
+
 ## 2.0.2
 
 - New: officer incidents